+++ programs/MushroomClassifier.php
@@ -8,9 +8,9 @@ use Rubix\ML\NeuralNet\Layers\Dense;
 use Rubix\ML\Reports\AggregateReport;
 use Rubix\ML\Reports\ConfusionMatrix;
 use Rubix\ML\Reports\PredictionSpeed;
-use Rubix\ML\NeuralNet\Layers\Dropout;
 use Rubix\ML\NeuralNet\Optimizers\Adam;
 use Rubix\ML\Transformers\OneHotEncoder;
+use Rubix\ML\NeuralNet\Layers\BatchNorm;
 use Rubix\ML\Reports\MulticlassBreakdown;
 use Rubix\ML\Transformers\LambdaFunction;
 use Rubix\ML\CrossValidation\Metrics\MCC;
@@ -47,9 +47,9 @@ $dataset = new Labeled($samples, $labels);
 
 $estimator = new Pipeline(new MultiLayerPerceptron([
     new Dense(30, new LeakyReLU()),
-    new Dropout(0.1),
+    new BatchNorm(),
     new Dense(30, new LeakyReLU()),
-    new Dropout(0.1),
+    new BatchNorm(),
     new Dense(30, new LeakyReLU()),
 ], 100, new Adam(0.001), 1e-4, new CrossEntropy(), 1e-3, new MCC(), 0.1, 3, 100), [
     new OneHotEncoder(),