+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - Unreleased
     - Added Gaussian Mixture clusterer
+    - Added Batch Norm hidden layer
     - Added Relative Entropy cost function to nn
     - Added random weighted subset to datasets
     - Committee Machine classifier only and added expert influence